Espada Eye lite/requirements.txt
torch==1.9.1
transformers==4.11.3
gradio==2.2.13
Pillow==8.2.0
